+++ src/components/PhoneNumberInput/index.js
@@ -1,7 +1,12 @@
 import { h } from 'preact'
 import PhoneNumber, {isValidPhoneNumber} from 'react-phone-number-input'
+import countries from 'react-phone-number-input/modules/countries'
+
 import classNames from 'classnames';
 import {localised} from '../../locales'
+import {lowerCase, upperCase} from '../utils/string'
+import {includes} from '../utils/array'
+
 
 import 'react-phone-number-input/rrui.css'
 import 'react-phone-number-input/style.css'
@@ -16,7 +21,7 @@ const FlagComponent = ({ countryCode, flagsPath }) => (
   />
 );
 
-const PhoneNumberInput = ({ translate, clearErrors, actions = {}}) => {
+const PhoneNumberInput = ({ translate, clearErrors, actions = {}, smsNumberCountryCode}) => {
 
   const onChange = (number) => {
     clearErrors()
@@ -24,11 +29,23 @@ const PhoneNumberInput = ({ translate, clearErrors, actions = {}}) => {
     actions.setMobileNumber({number, valid})
   }
 
+  const validCountryCode = () => {
+    const isCountryCodeInCountriesList = includes(countries.flat(), lowerCase(smsNumberCountryCode))
+    const isValid = smsNumberCountryCode.length === 2 && isCountryCodeInCountriesList
+    if (!isValid) { console.warn("Invalid ISO Country Code") }
+    return isValid
+  }
+
+  const countryCode = () => {
+    if (!smsNumberCountryCode) return 'GB'
+    return validCountryCode() ? upperCase(smsNumberCountryCode) : 'GB'
+  }
+
   return (
-    <form onSubmit={(e) => e.preventDefault()}>
+    <form onSubmit={(e) => e.preventDefault()} onError={(e) => console.log('shit happens', e)}>
       <PhoneNumber placeholder={translate('cross_device.phone_number_placeholder')}
         onChange={onChange}
-        country="GB"
+        country={countryCode()}
         inputClassName={`${style.mobileInput}`}
         className={`${style.phoneNumberContainer}`}
         flagComponent={ FlagComponent }